+++ mobile/android/android-components/docs/changelog.md
@@ -9,6 +9,11 @@ permalink: /changelog/
 * **browser-engine-gecko**
     * Added `WebExtensionInstallException.AdminInstallOnly` to handle the `ERROR_ADMIN_INSTALL_ONLY` error returned by Gecko when the add-on can only be installed via Enterprise Policies.
 
+* **feature-accounts-push**
+  * `CloseTabsUseCases.close()` now returns an `UndoableOperation`.
+  * 🆕 New `CloseTabsCommandReceiver` class for processing "close synced tabs" commands received from other devices.
+  * ⚠️ **Breaking change**: `CloseTabsFeature()` now takes a `CloseTabsCommandReceiver` instead of an `onTabsClosed` callback.
+
 # 128.0
 
 * **browser-toolbar**